requirements.txt
pyyaml
tinydb
click